moved the unit tests a bit
