+++ tests/lib.rs
@@ -1,10 +1,7 @@
 extern crate cdrs;
 extern crate byteorder;
+//we should write the integration tests over here
+//possibly in travis setup a cassandra node and try creating keyspace; creating table; inserting data into tables
+// reading from table
 
-mod frame;
-mod types;
 
-mod authenticators;
-mod compression;
-mod consistency;
-mod rows;
